#73 Handle getgauge module
- Install if not installed already
- Update if already installed
