+++ start.py
@@ -4,12 +4,10 @@ import platform
 import sys
 import threading
 from concurrent import futures
-from distutils import version
 from os import path
 
 
 import grpc
-import pkg_resources
 
 from getgauge import connection, processor
 from getgauge import lsp_server
@@ -17,7 +15,6 @@ from getgauge.impl_loader import copy_skel_files
 from getgauge.messages import lsp_pb2_grpc
 from getgauge.static_loader import load_files
 from getgauge.util import get_step_impl_dir
-from util import get_version
 
 PLUGIN_JSON = 'python.json'
 VERSION = 'version'
@@ -29,28 +26,8 @@ def main():
     if sys.argv[1] == "--init":
         copy_skel_files()
     else:
-        assert_versions()
         load_implementations()
         start()
-
-
-def assert_versions():
-    python_plugin_version = get_version()
-    getgauge_version = version.LooseVersion(pkg_resources.get_distribution('getgauge').version)
-    if (list(map(int, python_plugin_version.split(".")[0:3])) != getgauge_version.version[0:3]) or (
-            'dev' in getgauge_version.version and 'nightly' not in python_plugin_version) or (
-            'dev' not in getgauge_version.version and 'nightly' in python_plugin_version):
-        show_error_exit(python_plugin_version, getgauge_version)
-    if 'dev' in getgauge_version.version and 'nightly' in python_plugin_version:
-        if str(getgauge_version.version.pop()) not in python_plugin_version.replace("-", ""):
-            show_error_exit(python_plugin_version, getgauge_version)
-
-
-def show_error_exit(python_plugin_version, getgauge_version):
-    logging.fatal('Gauge-python({}) is not compatible with getgauge({}). Please install compatible versions.\n'.format(
-        python_plugin_version, getgauge_version))
-    exit(1)
-
 
 def load_implementations():
     d = get_step_impl_dir()
